proxy tests: compare the TTL ceiling in long arithmetic
A constant TTL ceiling large enough to overflow TTL * 1000 failed the build
(CS0220) instead of failing the anchoring and overflow tests at runtime, which
hid the guard from mutation testing.

+++ proxy/AI4IA.Proxy.Tests/AppConfigKeyPolicyTests.cs
@@ -129,7 +129,7 @@ public sealed class AppConfigKeyPolicyTests
         var deployed = new ProxyConfig();
         Assert.AreEqual(deployed.Timeout, AppConfigKeyPolicy.MaxTimeoutMs);
         Assert.AreEqual(deployed.DefaultTTLSecs, AppConfigKeyPolicy.MinTtlSecs);
-        Assert.AreEqual(deployed.Timeout, AppConfigKeyPolicy.MaxTtlSecs * 1000);
+        Assert.AreEqual((long)deployed.Timeout, AppConfigKeyPolicy.MaxTtlSecs * 1000L);
         Assert.IsTrue(AppConfigKeyPolicy.MinTimeoutMs < AppConfigKeyPolicy.MaxTimeoutMs);
 
         // Writing back the deployed values is always accepted.
